Add files via upload
The calculator task involves creating a basic calculator using HTML, CSS, and JavaScript. The calculator provides an interactive interface with buttons for addition, subtraction, multiplication, and division operations. It features a display screen to show user input and results. Utilizing CSS Grid for button alignments, the calculator uses event listeners, if-else statements, operators, and loops to handle user input and perform calculations. This project is designed for individuals with basic knowledge of HTML, CSS, and JavaScript, offering an opportunity to enhance skills in web development and user interaction.